requirements.txt
joblib==1.4.0
mlserver==1.5.0
mlserver-sklearn==0.5.2
numpy==1.24.2
pandas==1.5.2
scikit-learn==1.1.2
